♻️ [Refactor]: shareButton fsd 방식으로 리팩토링
