+++ src/features/share/ui/share.ui.types.ts
@@ -1,0 +1,3 @@
+export interface ShareButtonProps {
+  type: "copy" | "share";
+}
